+++ crates/merkle-tree/src/merkle_node.rs
@@ -17,7 +17,7 @@ pub enum InternalNode {
     /// A node that has not been fetched from storage yet.
     ///
     /// As such, all we know is its hash.
-    Unresolved(u32),
+    Unresolved(u64),
     /// A branch node with exactly two children.
     Binary(BinaryNode),
     /// Describes a path connecting two other nodes.
